Introduce server package
