+++ actors/akkachat/src/test/scala/com/nacionlumpen/server/CommandParserTest.scala
@@ -1,4 +1,4 @@
-package com.nacionlumpen
+package com.nacionlumpen.server
 
 import scalaz.DRight
